+++ tests/testthat/test-slice-assign.R
@@ -83,11 +83,13 @@ test_that("can assign lists", {
 
 test_that("atomics can't be assigned in lists", {
   x <- list(NULL)
-  expect_error(vec_slice(x, 1) <- NA, class = "vctrs_error_incompatible_type")
-  expect_error(vec_assign(x, 1, NA), class = "vctrs_error_incompatible_type")
-
   expect_error(vec_slice(x, 1) <- "foo", class = "vctrs_error_incompatible_type")
   expect_error(vec_assign(x, 1, "foo"), class = "vctrs_error_incompatible_type")
+
+  skip("FIXME: should we always allow assignment of unspecified vectors?")
+  # Should it then assign `NULL` rather than `NA`?
+  expect_error(vec_slice(x, 1) <- NA, class = "vctrs_error_incompatible_type")
+  expect_error(vec_assign(x, 1, NA), class = "vctrs_error_incompatible_type")
 })
 
 test_that("can assign and slice-assign data frames", {